# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: WanderP 旅行规划平台 API
  description: WanderP旅行规划平台后端API文档
  version: 1.0.0
  contact:
    name: WanderP Team
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
...
paths:
...
  /user/auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /user/auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /user/auth/refresh:
    post:
      tags: [认证]
      summary: 刷新令牌
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 刷新成功
  /user/auth/password:
    put:
      tags: [认证]
      summary: 修改密码
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: 修改成功

  # ==================== 用户接口 ====================
...
  /planner/plan/generate:
    post:
      tags: [AI规划]
      summary: AI生成行程规划
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination, days]
              properties:
                destination:
                  type: string
                  example: 东京
                days:
                  type: integer
                  minimum: 1
                  maximum: 30
                  example: 5
                preferences:
                  $ref: '#/components/schemas/Preference'
      responses:
        '200':
          description: 规划生成中
  /planner/plan/{id}/optimize:
    post:
      tags: [AI规划]
      summary: 优化行程规划
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                feedback:
                  type: string
      responses:
        '200':
          description: 优化中
  /planner/plan/{id}/activities:
    post:
      tags: [AI规划]
      summary: 添加活动
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        '200':
          description: 添加成功
  /planner/plan/activities/{activityId}/vote:
    post:
      tags: [AI规划]
      summary: 活动投票
      security:
        - BearerAuth: []
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [up, down]
      responses:
        '200':
          description: 投票成功
  /planner/plan/recommendations:
    get:
      tags: [AI规划]
      summary: 获取推荐活动
      parameters:
        - name: destination
          in: query
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 推荐列表

  # ==================== 偏好接口 ====================
  /planner/preferences:
    get:
      tags: [偏好]
      summary: 获取用户偏好
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 偏好设置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preference'
    put:
      tags: [偏好]
      summary: 更新用户偏好
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Preference'
      responses:
        '200':
          description: 更新成功

  # ==================== 路线规划接口 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Response:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            token:
              type: string
            expires_at:
              type: integer
            user:
              $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nickname:
          type: string
        avatar:
          type: string
        bio:
          type: string
        gender:
          type: string
          enum: [male, female, unknown]
        location:
          type: string
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        trip_id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [sightseeing, food, transport, accommodation, shopping, entertainment]
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
        cost:
          type: number
          format: double
        sort_order:
          type: integer
        day_index:
          type: integer
...
    Preference:
      type: object
      properties:
        travel_style:
          type: string
          enum: [relax, adventure, culture, food, nature]
        taste_prefs:
          type: string
        budget_level:
          type: string
          enum: [low, medium, high]
        pace_level:
          type: string
          enum: [slow, moderate, fast]
        accommodation:
          type: string
          enum: [hotel, hostel, apartment, camping]
        transport:
          type: string
          enum: [walking, driving, public, ride]
        interests:
          type: array
          items:
            type: string
        avoidances:
          type: array
          items:
            type: string
...